1078-查询区域或线路数据应答...

+++ src/main/java/com/genersoft/iot/vmp/jt1078/proc/response/J8608.java
@@ -30,10 +30,10 @@ public class J8608 extends Rs {
         ByteBuf buffer = Unpooled.buffer();
         buffer.writeByte(type);
         if (idList == null || idList.isEmpty()) {
-            buffer.writeByte(0);
+            buffer.writeInt(0);
             return buffer;
         }else {
-            buffer.writeByte(idList.size());
+            buffer.writeInt(idList.size());
         }
         for (Long id : idList) {
             buffer.writeInt((int) (id & 0xffffffffL));
